+++ cargo-embed/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update to latest probe-rs
+
 ### Fixed
 
 ## [0.12.0]